Fix error in code, now working
There was an error where code was run two times and the second time on
the resulting dataframe which then through an error as the variables
vere not present.
Also added a `possibly()` to catch and continue with the function if
there are no resulting data, i.e. an empty dataframe.